refactor support code and add langchain wrapper
