+++ api/terraform/python/openai_text/const.py
@@ -1,0 +1,43 @@
+import os
+import openai
+
+HTTP_RESPONSE_OK = 200
+HTTP_RESPONSE_BAD_REQUEST = 400
+HTTP_RESPONSE_INTERNAL_SERVER_ERROR = 500
+DEBUG_MODE = os.getenv("DEBUG_MODE", "False").lower() in ("true", "1", "t")
+
+
+class OpenAIEndPoint:
+    """
+    A class representing an endpoint for the OpenAI API.
+
+    Attributes:
+        api_key (str): The API key to use for authentication.
+        endpoint (str): The URL of the OpenAI API endpoint.
+    """
+
+    Embedding = openai.Embedding.__name__
+    ChatCompletion = openai.ChatCompletion.__name__
+    Moderation = openai.Moderation.__name__
+    Image = openai.Image.__name__
+    Audio = openai.Audio.__name__
+    Models = openai.Model.__name__
+    all_endpoints = [Embedding, ChatCompletion, Moderation, Image, Audio, Models]
+
+
+VALID_CHAT_COMPLETION_MODELS = [
+    "gpt-4",
+    "gpt-4-0613",
+    "gpt-4-32k",
+    "gpt-4-32k-0613",
+    "gpt-3.5-turbo",
+    "gpt-3.5-turbo-0613",
+    "gpt-3.5-turbo-16k",
+    "gpt-3.5-turbo-16k-0613",
+]
+VALID_EMBEDDING_MODELS = [
+    "text-embedding-ada-002",
+    "text-similarity-*-001",
+    "text-search-*-*-001",
+    "code-search-*-*-001",
+]
