+++ api/terraform/python/openai_text/data/example_request.json
@@ -1,0 +1,14 @@
+{
+  "model": "gpt-3.5-turbo",
+  "end_point": "ChatCompletion",
+  "messages": [
+      {
+          "role": "system",
+          "content": "You will be provided with statements, and your task is to convert them to standard English."
+      },
+      {
+          "role": "user",
+          "content": "She no went to the market."
+      }
+  ]
+}
