+++ api/terraform/python/openai_text/tests/test_init.py
@@ -1,7 +1,17 @@
 from dotenv import load_dotenv, find_dotenv
 import os
 import json
-from openai_text import handler
+from openai_text.openai_text import handler
+
+# Load environment variables from .env file in all folders
+dotenv_path = find_dotenv()
+if os.path.exists(dotenv_path):
+    load_dotenv(dotenv_path=dotenv_path, verbose=True)
+    OPENAI_API_KEY = os.environ["OPENAI_API_KEY"]
+    OPENAI_API_ORGANIZATION = os.environ["OPENAI_API_ORGANIZATION"]
+    PINECONE_API_KEY = os.environ["PINECONE_API_KEY"]
+else:
+    raise Exception("No .env file found in root directory of repository")
 
 
 def get_event(filespec):
@@ -19,14 +29,3 @@ def handle_event(event):
         pinecone_api_key=PINECONE_API_KEY,
     )
     return retval
-
-
-# Load environment variables from .env file in all folders
-dotenv_path = find_dotenv()
-if os.path.exists(dotenv_path):
-    load_dotenv(dotenv_path=dotenv_path, verbose=True)
-    OPENAI_API_KEY = os.environ["OPENAI_API_KEY"]
-    OPENAI_API_ORGANIZATION = os.environ["OPENAI_API_ORGANIZATION"]
-    PINECONE_API_KEY = os.environ["PINECONE_API_KEY"]
-else:
-    raise Exception("No .env file found in root directory of repository")
